Fix only_claiming_mints_tokens rule
The calls to `get_total_supply` were implicitly assuming that the token was registered.  This creates a vacuity when the target function is `register_foreign_token`, which assumes that the token is *not* registered.

With this change, combined with https://github.com/Certora/cvl-move-proto/pull/7, the rule now passes for all targets.